moving high-level-synthesis plugins to a separate file
