+++ qiskit/transpiler/passes/synthesis/high_level_synthesis.py
@@ -10,153 +10,8 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-
 """
-
-High Level Synthesis Plugins
------------------------------
-
-Clifford Synthesis
-''''''''''''''''''
-
-.. list-table:: Plugins for :class:`qiskit.quantum_info.Clifford` (key = ``"clifford"``)
-    :header-rows: 1
-
-    * - Plugin name
-      - Plugin class
-      - Targeted connectivity
-      - Description
-    * - ``"ag"``
-      - :class:`~.AGSynthesisClifford`
-      - all-to-all
-      - greedily optimizes CX-count
-    * - ``"bm"``
-      - :class:`~.BMSynthesisClifford`
-      - all-to-all
-      - optimal count for `n=2,3`; used in ``"default"`` for `n=2,3`
-    * - ``"greedy"``
-      - :class:`~.GreedySynthesisClifford`
-      - all-to-all
-      - greedily optimizes CX-count; used in ``"default"`` for `n>=4`
-    * - ``"layers"``
-      - :class:`~.LayerSynthesisClifford`
-      - all-to-all
-      -
-    * - ``"lnn"``
-      - :class:`~.LayerLnnSynthesisClifford`
-      - linear
-      - many CX-gates but guarantees CX-depth of at most `7*n+2`
-    * - ``"default"``
-      - :class:`~.DefaultSynthesisClifford`
-      - all-to-all
-      - usually best for optimizing CX-count (and optimal CX-count for `n=2,3`)
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   AGSynthesisClifford
-   BMSynthesisClifford
-   GreedySynthesisClifford
-   LayerSynthesisClifford
-   LayerLnnSynthesisClifford
-   DefaultSynthesisClifford
-
-
-Linear Function Synthesis
-'''''''''''''''''''''''''
-
-.. list-table:: Plugins for :class:`.LinearFunction` (key = ``"linear"``)
-    :header-rows: 1
-
-    * - Plugin name
-      - Plugin class
-      - Targeted connectivity
-      - Description
-    * - ``"kms"``
-      - :class:`~.KMSSynthesisLinearFunction`
-      - linear
-      - many CX-gates but guarantees CX-depth of at most `5*n`
-    * - ``"pmh"``
-      - :class:`~.PMHSynthesisLinearFunction`
-      - all-to-all
-      - greedily optimizes CX-count; used in ``"default"``
-    * - ``"default"``
-      - :class:`~.DefaultSynthesisLinearFunction`
-      - all-to-all
-      - best for optimizing CX-count
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   KMSSynthesisLinearFunction
-   PMHSynthesisLinearFunction
-   DefaultSynthesisLinearFunction
-
-
-Permutation Synthesis
-'''''''''''''''''''''
-
-.. list-table:: Plugins for :class:`.PermutationGate` (key = ``"permutation"``)
-    :header-rows: 1
-
-    * - Plugin name
-      - Plugin class
-      - Targeted connectivity
-      - Description
-    * - ``"basic"``
-      - :class:`~.BasicSynthesisPermutation`
-      - all-to-all
-      - optimal SWAP-count; used in ``"default"``
-    * - ``"acg"``
-      - :class:`~.ACGSynthesisPermutation`
-      - all-to-all
-      - guarantees SWAP-depth of at most `2`
-    * - ``"kms"``
-      - :class:`~.KMSSynthesisPermutation`
-      - linear
-      - many SWAP-gates, but guarantees SWAP-depth of at most `n`
-    * - ``"token_swapper"``
-      - :class:`~.TokenSwapperSynthesisPermutation`
-      - any
-      - greedily optimizes SWAP-count for arbitrary connectivity
-    * - ``"default"``
-      - :class:`~.BasicSynthesisPermutation`
-      - all-to-all
-      - best for optimizing SWAP-count
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   BasicSynthesisPermutation
-   ACGSynthesisPermutation
-   KMSSynthesisPermutation
-   TokenSwapperSynthesisPermutation
-
-
-QFT Synthesis
-'''''''''''''
-
-.. list-table:: Plugins for :class:`.QFTGate` (key = ``"qft"``)
-    :header-rows: 1
-
-    * - Plugin name
-      - Plugin class
-      - Targeted connectivity
-    * - ``"full"``
-      - :class:`~.QFTSynthesisFull`
-      - all-to-all
-    * - ``"line"``
-      - :class:`~.QFTSynthesisLine`
-      - linear
-    * - ``"default"``
-      - :class:`~.QFTSynthesisFull`
-      - all-to-all
-
-.. autosummary::
-   :toctree: ../stubs/
-
-   QFTSynthesisFull
-   QFTSynthesisLine
+High-level-synthesis transpiler pass.
 """
 
 from __future__ import annotations
@@ -165,20 +20,17 @@ import typing
 from typing import Optional, Union, List, Tuple, Callable, Sequence
 
 import numpy as np
-import rustworkx as rx
 
 from qiskit.circuit.operation import Operation
 from qiskit.converters import circuit_to_dag, dag_to_circuit
 from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.circuit.quantumcircuit import QuantumCircuit
 from qiskit.circuit import ControlledGate, EquivalenceLibrary, equivalence
-from qiskit.circuit.library import LinearFunction
 from qiskit.transpiler.passes.utils import control_flow
 from qiskit.transpiler.target import Target
 from qiskit.transpiler.coupling import CouplingMap
 from qiskit.dagcircuit.dagcircuit import DAGCircuit
 from qiskit.transpiler.exceptions import TranspilerError
-from qiskit.transpiler.passes.routing.algorithms import ApproximateTokenSwapper
 
 from qiskit.circuit.annotated_operation import (
     AnnotatedOperation,
@@ -186,32 +38,8 @@ from qiskit.circuit.annotated_operation import (
     ControlModifier,
     PowerModifier,
 )
-from qiskit.circuit.library import QFTGate
-from qiskit.synthesis.clifford import (
-    synth_clifford_full,
-    synth_clifford_layers,
-    synth_clifford_depth_lnn,
-    synth_clifford_greedy,
-    synth_clifford_ag,
-    synth_clifford_bm,
-)
-from qiskit.synthesis.linear import (
-    synth_cnot_count_full_pmh,
-    synth_cnot_depth_line_kms,
-    calc_inverse_matrix,
-)
-from qiskit.synthesis.linear.linear_circuits_utils import transpose_cx_circ
-from qiskit.synthesis.permutation import (
-    synth_permutation_basic,
-    synth_permutation_acg,
-    synth_permutation_depth_lnn_kms,
-)
-from qiskit.synthesis.qft import (
-    synth_qft_full,
-    synth_qft_line,
-)
 
-from .plugin import HighLevelSynthesisPluginManager, HighLevelSynthesisPlugin
+from .plugin import HighLevelSynthesisPluginManager
 
 if typing.TYPE_CHECKING:
     from qiskit.dagcircuit import DAGOpNode
@@ -718,422 +546,4 @@ class HighLevelSynthesis(TransformationPass):
                     raise TranspilerError(f"Unknown modifier {modifier}.")
 
             return synthesized_op
-        return None
-
-
-class DefaultSynthesisClifford(HighLevelSynthesisPlugin):
-    """The default clifford synthesis plugin.
-
-    For N <= 3 qubits this is the optimal CX cost decomposition by Bravyi, Maslov.
-    For N > 3 qubits this is done using the general non-optimal greedy compilation
-    routine from reference by Bravyi, Hu, Maslov, Shaydulin.
-
-    This plugin name is :``clifford.default`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Clifford."""
-        decomposition = synth_clifford_full(high_level_object)
-        return decomposition
-
-
-class AGSynthesisClifford(HighLevelSynthesisPlugin):
-    """Clifford synthesis plugin based on the Aaronson-Gottesman method.
-
-    This plugin name is :``clifford.ag`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Clifford."""
-        decomposition = synth_clifford_ag(high_level_object)
-        return decomposition
-
-
-class BMSynthesisClifford(HighLevelSynthesisPlugin):
-    """Clifford synthesis plugin based on the Bravyi-Maslov method.
-
-    The method only works on Cliffords with at most 3 qubits, for which it
-    constructs the optimal CX cost decomposition.
-
-    This plugin name is :``clifford.bm`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Clifford."""
-        if high_level_object.num_qubits <= 3:
-            decomposition = synth_clifford_bm(high_level_object)
-        else:
-            decomposition = None
-        return decomposition
-
-
-class GreedySynthesisClifford(HighLevelSynthesisPlugin):
-    """Clifford synthesis plugin based on the greedy synthesis
-    Bravyi-Hu-Maslov-Shaydulin method.
-
-    This plugin name is :``clifford.greedy`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Clifford."""
-        decomposition = synth_clifford_greedy(high_level_object)
-        return decomposition
-
-
-class LayerSynthesisClifford(HighLevelSynthesisPlugin):
-    """Clifford synthesis plugin based on the Bravyi-Maslov method
-    to synthesize Cliffords into layers.
-
-    This plugin name is :``clifford.layers`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Clifford."""
-        decomposition = synth_clifford_layers(high_level_object)
-        return decomposition
-
-
-class LayerLnnSynthesisClifford(HighLevelSynthesisPlugin):
-    """Clifford synthesis plugin based on the Bravyi-Maslov method
-    to synthesize Cliffords into layers, with each layer synthesized
-    adhering to LNN connectivity.
-
-    This plugin name is :``clifford.lnn`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Clifford."""
-        decomposition = synth_clifford_depth_lnn(high_level_object)
-        return decomposition
-
-
-class DefaultSynthesisLinearFunction(HighLevelSynthesisPlugin):
-    """The default linear function synthesis plugin.
-
-    This plugin name is :``linear_function.default`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given LinearFunction."""
-        decomposition = synth_cnot_count_full_pmh(high_level_object.linear)
-        return decomposition
-
-
-class KMSSynthesisLinearFunction(HighLevelSynthesisPlugin):
-    """Linear function synthesis plugin based on the Kutin-Moulton-Smithline method.
-
-    This plugin name is :``linear_function.kms`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-
-    The plugin supports the following plugin-specific options:
-
-    * use_inverted: Indicates whether to run the algorithm on the inverse matrix
-        and to invert the synthesized circuit.
-        In certain cases this provides a better decomposition than the direct approach.
-    * use_transposed: Indicates whether to run the algorithm on the transposed matrix
-        and to invert the order of CX gates in the synthesized circuit.
-        In certain cases this provides a better decomposition than the direct approach.
-
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given LinearFunction."""
-
-        if not isinstance(high_level_object, LinearFunction):
-            raise TranspilerError(
-                "PMHSynthesisLinearFunction only accepts objects of type LinearFunction"
-            )
-
-        use_inverted = options.get("use_inverted", False)
-        use_transposed = options.get("use_transposed", False)
-
-        mat = high_level_object.linear.astype(bool, copy=False)
-
-        if use_transposed:
-            mat = np.transpose(mat)
-        if use_inverted:
-            mat = calc_inverse_matrix(mat)
-
-        decomposition = synth_cnot_depth_line_kms(mat)
-
-        if use_transposed:
-            decomposition = transpose_cx_circ(decomposition)
-        if use_inverted:
-            decomposition = decomposition.inverse()
-
-        return decomposition
-
-
-class PMHSynthesisLinearFunction(HighLevelSynthesisPlugin):
-    """Linear function synthesis plugin based on the Patel-Markov-Hayes method.
-
-    This plugin name is :``linear_function.pmh`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-
-    The plugin supports the following plugin-specific options:
-
-    * section size: The size of each section used in the Patel–Markov–Hayes algorithm [1].
-    * use_inverted: Indicates whether to run the algorithm on the inverse matrix
-        and to invert the synthesized circuit.
-        In certain cases this provides a better decomposition than the direct approach.
-    * use_transposed: Indicates whether to run the algorithm on the transposed matrix
-        and to invert the order of CX gates in the synthesized circuit.
-        In certain cases this provides a better decomposition than the direct approach.
-
-    References:
-        1. Patel, Ketan N., Igor L. Markov, and John P. Hayes,
-           *Optimal synthesis of linear reversible circuits*,
-           Quantum Information & Computation 8.3 (2008): 282-294.
-           `arXiv:quant-ph/0302002 [quant-ph] <https://arxiv.org/abs/quant-ph/0302002>`_
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given LinearFunction."""
-
-        if not isinstance(high_level_object, LinearFunction):
-            raise TranspilerError(
-                "PMHSynthesisLinearFunction only accepts objects of type LinearFunction"
-            )
-
-        section_size = options.get("section_size", 2)
-        use_inverted = options.get("use_inverted", False)
-        use_transposed = options.get("use_transposed", False)
-
-        mat = high_level_object.linear.astype(bool, copy=False)
-
-        if use_transposed:
-            mat = np.transpose(mat)
-        if use_inverted:
-            mat = calc_inverse_matrix(mat)
-
-        decomposition = synth_cnot_count_full_pmh(mat, section_size=section_size)
-
-        if use_transposed:
-            decomposition = transpose_cx_circ(decomposition)
-        if use_inverted:
-            decomposition = decomposition.inverse()
-
-        return decomposition
-
-
-class KMSSynthesisPermutation(HighLevelSynthesisPlugin):
-    """The permutation synthesis plugin based on the Kutin, Moulton, Smithline method.
-
-    This plugin name is :``permutation.kms`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Permutation."""
-        decomposition = synth_permutation_depth_lnn_kms(high_level_object.pattern)
-        return decomposition
-
-
-class BasicSynthesisPermutation(HighLevelSynthesisPlugin):
-    """The permutation synthesis plugin based on sorting.
-
-    This plugin name is :``permutation.basic`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Permutation."""
-        decomposition = synth_permutation_basic(high_level_object.pattern)
-        return decomposition
-
-
-class ACGSynthesisPermutation(HighLevelSynthesisPlugin):
-    """The permutation synthesis plugin based on the Alon, Chung, Graham method.
-
-    This plugin name is :``permutation.acg`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Permutation."""
-        decomposition = synth_permutation_acg(high_level_object.pattern)
-        return decomposition
-
-
-class QFTSynthesisFull(HighLevelSynthesisPlugin):
-    """Synthesis plugin for QFT gates using all-to-all connectivity.
-
-    This plugin name is :``qft.full`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-
-    The plugin supports the following additional options:
-
-    * reverse_qubits (bool): Whether to synthesize the "QFT" operation (if ``False``,
-        which is the default) or the "QFT-with-reversal" operation (if ``True``).
-        Some implementation of the ``QFTGate`` include a layer of swap gates at the end
-        of the synthesized circuit, which can in principle be dropped if the ``QFTGate``
-        itself is the last gate in the circuit.
-    * approximation_degree (int): The degree of approximation (0 for no approximation).
-        It is possible to implement the QFT approximately by ignoring
-        controlled-phase rotations with the angle beneath a threshold. This is discussed
-        in more detail in [1] or [2].
-    * insert_barriers (bool): If True, barriers are inserted as visualization improvement.
-    * inverse (bool): If True, the inverse Fourier transform is constructed.
-    * name (str): The name of the circuit.
-
-    References:
-        1. Adriano Barenco, Artur Ekert, Kalle-Antti Suominen, and Päivi Törmä,
-           *Approximate Quantum Fourier Transform and Decoherence*,
-           Physical Review A (1996).
-           `arXiv:quant-ph/9601018 [quant-ph] <https://arxiv.org/abs/quant-ph/9601018>`_
-        2. Donny Cheung,
-           *Improved Bounds for the Approximate QFT* (2004),
-           `arXiv:quant-ph/0403071 [quant-ph] <https://https://arxiv.org/abs/quant-ph/0403071>`_
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given QFTGate."""
-        if not isinstance(high_level_object, QFTGate):
-            raise TranspilerError(
-                "The synthesis plugin 'qft.full` only applies to objects of type QFTGate."
-            )
-
-        reverse_qubits = options.get("reverse_qubits", False)
-        approximation_degree = options.get("approximation_degree", 0)
-        insert_barriers = options.get("insert_barriers", False)
-        inverse = options.get("inverse", False)
-        name = options.get("name", None)
-
-        decomposition = synth_qft_full(
-            num_qubits=high_level_object.num_qubits,
-            do_swaps=not reverse_qubits,
-            approximation_degree=approximation_degree,
-            insert_barriers=insert_barriers,
-            inverse=inverse,
-            name=name,
-        )
-        return decomposition
-
-
-class QFTSynthesisLine(HighLevelSynthesisPlugin):
-    """Synthesis plugin for QFT gates using linear connectivity.
-
-    This plugin name is :``qft.line`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-
-    The plugin supports the following additional options:
-
-    * reverse_qubits (bool): Whether to synthesize the "QFT" operation (if ``False``,
-        which is the default) or the "QFT-with-reversal" operation (if ``True``).
-        Some implementation of the ``QFTGate`` include a layer of swap gates at the end
-        of the synthesized circuit, which can in principle be dropped if the ``QFTGate``
-        itself is the last gate in the circuit.
-    * approximation_degree (int): the degree of approximation (0 for no approximation).
-        It is possible to implement the QFT approximately by ignoring
-        controlled-phase rotations with the angle beneath a threshold. This is discussed
-        in more detail in [1] or [2].
-
-    References:
-        1. Adriano Barenco, Artur Ekert, Kalle-Antti Suominen, and Päivi Törmä,
-           *Approximate Quantum Fourier Transform and Decoherence*,
-           Physical Review A (1996).
-           `arXiv:quant-ph/9601018 [quant-ph] <https://arxiv.org/abs/quant-ph/9601018>`_
-        2. Donny Cheung,
-           *Improved Bounds for the Approximate QFT* (2004),
-           `arXiv:quant-ph/0403071 [quant-ph] <https://https://arxiv.org/abs/quant-ph/0403071>`_
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given QFTGate."""
-        if not isinstance(high_level_object, QFTGate):
-            raise TranspilerError(
-                "The synthesis plugin 'qft.line` only applies to objects of type QFTGate."
-            )
-
-        reverse_qubits = options.get("reverse_qubits", False)
-        approximation_degree = options.get("approximation_degree", 0)
-
-        decomposition = synth_qft_line(
-            num_qubits=high_level_object.num_qubits,
-            do_swaps=not reverse_qubits,
-            approximation_degree=approximation_degree,
-        )
-        return decomposition
-
-
-class TokenSwapperSynthesisPermutation(HighLevelSynthesisPlugin):
-    """The permutation synthesis plugin based on the token swapper algorithm.
-
-    This plugin name is :``permutation.token_swapper`` which can be used as the key on
-    an :class:`~.HLSConfig` object to use this method with :class:`~.HighLevelSynthesis`.
-
-    In more detail, this plugin is used to synthesize objects of type `PermutationGate`.
-    When synthesis succeeds, the plugin outputs a quantum circuit consisting only of swap
-    gates. When synthesis does not succeed, the plugin outputs `None`.
-
-    If either `coupling_map` or `qubits` is None, then the synthesized circuit
-    is not required to adhere to connectivity constraints, as is the case
-    when the synthesis is done before layout/routing.
-
-    On the other hand, if both `coupling_map` and `qubits` are specified, the synthesized
-    circuit is supposed to adhere to connectivity constraints. At the moment, the
-    plugin only creates swap gates between qubits in `qubits`, i.e. it does not use
-    any other qubits in the coupling map (if such synthesis is not possible, the
-    plugin  outputs `None`).
-
-    The plugin supports the following plugin-specific options:
-
-    * trials: The number of trials for the token swapper to perform the mapping. The
-      circuit with the smallest number of SWAPs is returned.
-    * seed: The argument to the token swapper specifying the seed for random trials.
-    * parallel_threshold: The argument to the token swapper specifying the number of nodes
-      in the graph beyond which the algorithm will use parallel processing.
-
-    For more details on the token swapper algorithm, see to the paper:
-    `arXiv:1902.09102 <https://arxiv.org/abs/1902.09102>`__.
-    """
-
-    def run(self, high_level_object, coupling_map=None, target=None, qubits=None, **options):
-        """Run synthesis for the given Permutation."""
-
-        trials = options.get("trials", 5)
-        seed = options.get("seed", 0)
-        parallel_threshold = options.get("parallel_threshold", 50)
-
-        pattern = high_level_object.pattern
-        pattern_as_dict = {j: i for i, j in enumerate(pattern)}
-
-        # When the plugin is called from the HighLevelSynthesis transpiler pass,
-        # the coupling map already takes target into account.
-        if coupling_map is None or qubits is None:
-            # The abstract synthesis uses a fully connected coupling map, allowing
-            # arbitrary connections between qubits.
-            used_coupling_map = CouplingMap.from_full(len(pattern))
-        else:
-            # The concrete synthesis uses the coupling map restricted to the set of
-            # qubits over which the permutation gate is defined. If we allow using other
-            # qubits in the coupling map, replacing the node in the DAGCircuit that
-            # defines this PermutationGate by the DAG corresponding to the constructed
-            # decomposition becomes problematic. Note that we allow the reduced
-            # coupling map to be disconnected.
-            used_coupling_map = coupling_map.reduce(qubits, check_if_connected=False)
-
-        graph = used_coupling_map.graph.to_undirected()
-        swapper = ApproximateTokenSwapper(graph, seed=seed)
-
-        try:
-            swapper_result = swapper.map(
-                pattern_as_dict, trials, parallel_threshold=parallel_threshold
-            )
-        except rx.InvalidMapping:
-            swapper_result = None
-
-        if swapper_result is not None:
-            decomposition = QuantumCircuit(len(graph.node_indices()))
-            for swap in swapper_result:
-                decomposition.swap(*swap)
-            return decomposition
-
         return None
